refactor: SearchCardを目的別ファイルに分割しUI stateとクエリstateを分離
832行のモノリシックな SearchCard.tsx を以下へ分割する。

- types.ts: 型定義
- searchQueryUtils.ts: クエリ文字列組み立て等の純粋関数
- useSearchCardQuery.ts: 検索条件（API クエリ params）state を一元管理するフック
- QuickSearchDropdown.tsx / QuickSearchButtons.tsx: クイック検索UI
- SearchFieldsGrid.tsx: 検索フィールドのグリッドレイアウト
- CalendarDateButton.tsx / DatePeriodBlock.tsx: 日付検索UI
- SearchCard.tsx: 展開/折りたたみの UI 表示 state のみを持つ本体（832→281行）

展開状態（isExpanded）は UI 表示 state として SearchCard 本体に残し、
検索条件（selectedQueries / dateSegment / dateInputs / isYearPickerOpen）は
useSearchCardQuery フックへ切り出すことで、
「UI 表示 state」と「API query params state」を分離した。

公開 API（SearchCard の props）は変更していないため、
呼び出し側・テストへの影響はない。
